backend/requirements_update.txt
flask
scikit-learn
joblib
deepface
opencv-python
transformers
torch
google-api-python-client
httplib2
requests
